[fix] sep msgbs file .
