+++ go/pkg/msgbs/router.go
@@ -1,0 +1,11 @@
+package msgbs
+
+type Router map[Event][]SubscribeFunc
+
+func NewRouter() Router {
+	return Router{}
+}
+
+func (r Router) Subscribe(evnt Event, subsc SubscribeFunc) {
+	r[evnt] = append(r[evnt], subsc)
+}
